+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes
 
-## 0.6.5
+## UNRELEASED
 
 - Implement CLI workflow. User will use script `spectacle-cli init` to kick off the workflow to create a server or build directory.